Move content out of App
