+++ client/src/components/ChatScreen/index.js
@@ -1,0 +1,3 @@
+import ChatScreen from './ChatScreen';
+
+export default ChatScreen;
